mcmmo compatibility is here
if mcMMO is enabled on the server
the stats of the claimed horse will be based on the player's taming level
and fightning while riding a horse will give you xp in taming

Fixed a little exploit where you could get an infinite supply of saddles and armours.
Fixed an issue with my code.
Refactored a piece of code